dom/bindings/test/test_Object.prototype_props.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test for bug 987110</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
/* global test, assert_array_equals */
test(function() {
  var props = Object.getOwnPropertyNames(Object.prototype);
  // If you change this list, make sure it continues to match the list in
  // Codegen.py's CGDictionary.getMemberDefinition method.
  var expected = [
      "constructor", "toSource", "toString", "toLocaleString", "valueOf",
      "hasOwnProperty", "isPrototypeOf", "propertyIsEnumerable",
      "__defineGetter__", "__defineSetter__", "__lookupGetter__",
      "__lookupSetter__", "__proto__",
    ];
  assert_array_equals(props.sort(), expected.sort());
}, "Own properties of Object.prototype");
</script>
